[Refactor] console view와 관련된 클래스 패키징
